Fix textwrapping on reference and design examples
The code examples have text in comments that wrap to the next line. The
result is text that looks misleading and confusing by having a single
word stick out to the next line.

Similar to other long, multi-line comments this moves words or phrases
to the next line to prevent the wrapping.

Fixes #24